Updated migrations, testing, models, views
Updated to make use of new weather_current table, this table contains 1 row (the latest weather) and will be updated frequently on a cronjob with update_weather_current.py
As django tests are designed to run in isolation, a test database must be setup (the structure mirrors our production database, but with junk data instead).
Note the sql script included is only used to create the test database, which is torn down after testing is complete.